Restructured folder into submodules and improved typing by adding generics
