+++ tests/test_regularization.py
@@ -1,0 +1,6 @@
+"""Tests for `imitation.regularization.*`."""
+
+import numpy as np
+import pytest
+
+from imitation import regularization
